1: two sum
give an array of integers and a target integer, return the indices of two numbers that add up to the target.

no negative numbers, order does not matter. dont use the same element twice.

use a hashmap, mapping a number to its index in order to find the required value and index.

time complexity O(n), because average case of hashmap lookup is O(1)

space complexity: O(n) number of elements in the hashmap