21: merge two sorted list
start with a dummy head node and a current pointer for the building result.

loo when list1 and and list2 are not empty.
append lower node to cursor of dummy node and remove node from the list.
finally append the list that is not empty.

Time Complexity: O(n + m) where n and m are the length of both lists. Each node is processed onces.

Space Complexity: O(1)